#1 Added more settings to cropper object
Standard view mode is set to "the crop box should be within the canvas" to avoid that user can crop photo outside canvas, which would generate white spots around image.
Drag mode is set to "none" so user can not create a new crop box or move the canvas around.
Drag mode on double click is disabled to avoid that user can use that functionality to drag canvas around (as drag mode does not solve that alone)