Refactored gallery functionality to reduce complexity in main state
Gallery work better as a standalone component with internal state, rather than a global state.
This make it possible to navigate through tent sites in the different views (explore, bookmarks, contributions).

Refactored all use of gallery and tent sites to make it work.
Also refactored how bookmark info is fetched, so it know which users who has bookmarked, not only the count.

Removed PhotoWall as Explore.vue implement all the functionality - should not be a component to avoid that all content structured alike.