Refactored state of tent sites and fixed bug where comments were not loaded properly
We need to share state of tent sites regarding comments, bookmarks etc. between different views (locate, explore, bookmarks, contributions).
Now all components dispatch tent sites to a global state, and keep track of ids for tent sites that are to be displayed on that component internally.
By doing that it is possible to share state, without displaying tent sites that only are to be displayed in the given component.

This also covered up a bug where comments were not loaded. Though there is still a problem where comments are not displayed right away.